backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
anthropic==0.18.1
openai==1.12.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.4
